chapters/ch4/ep4/python/ex2/requirements.txt
pandas==1.3.5
jupyterlab==3.4.2
Flask==2.1.2
PyYAML==6.0
PyMySQL==1.0.2
SQLAlchemy==1.4.36
google-cloud-bigquery==3.1.0
gcsfs==2022.3.0
pandas-gbq==0.17.4
pyarrow==7.0.0
